Initial setup - controller, service, and restaurant data container.
Setup a basic controller for testing.
ZomatoService is a first pass at the needed service, but needs to be tested since it is a first attempt at using custom headers for a ResponseEntity.

Restaurant object is a simple data container for displaying data about a restaurant - the data it contains can be changed, but basic data was chosen for testing purposes.

JUnit tests will be implemented next.