Added restaurant id and frame for JUnit tests.
Restaurant id added to the Restaurant data class.

The keys for accessing the JSON fields returned by the api have been updated to reflect a missed outer key.

Start of the code that will be used for testing the ZomatoService and added the file that will be used for testing the FrontEndController.

Next is filling in the rest of the test code and adding additional tests for extra coverage.