Controller and Service work. Removed unnecessary test files.
ZomatoService logic tested in Postman and works.

Controller was tested with a basic view and it receives the expected result from the Service.

Test files were removed as the ZomatoService required an end-to-end test (which was performed in Postman), much like the WeatherService in our past assignments.
The FrontEndController (at this point) only receives the returned value from the ZomatoService and passes it to a View (html file), so no functional logic present.